Fix
roundf to round

+++ prog.c
@@ -6,7 +6,7 @@ int main(void) {
   scanf("%f", &(len));
   if (len>0){
   //printf("%f\n", len);
-  printf("%g", 0.001*roundf(1000*len*len*sqrt(3))/4);
+  printf("%g", 0.001*round(1000*len*len*sqrt(3))/4);
   return 0;
   }else{return 1;}
 }
